+++ docker-qgis/process_projectfile.py
@@ -1,8 +1,5 @@
-import html
-import io
 import logging
 from pathlib import Path
-from typing import NamedTuple, Optional
 from xml.etree import ElementTree
 
 from qfieldcloud.qgis.utils import (
@@ -11,6 +8,7 @@ from qfieldcloud.qgis.utils import (
     InvalidXmlFileException,
     ProjectFileNotFoundException,
     get_layers_data,
+    get_qgis_xml_error_context,
     layers_data_to_string,
 )
 from qgis.core import QgsMapRendererParallelJob, QgsMapSettings, QgsProject
@@ -18,50 +16,6 @@ from qgis.PyQt.QtCore import QEventLoop, QSize
 from qgis.PyQt.QtGui import QColor
 
 logger = logging.getLogger("PROCPRJ")
-
-
-class XmlErrorLocation(NamedTuple):
-    line: int
-    column: int
-
-
-def get_qgis_xml_error_location(
-    invalid_token_error_msg: str,
-) -> Optional[XmlErrorLocation]:
-    """Get column and line numbers from the provided error message."""
-    if "invalid token" not in invalid_token_error_msg.casefold():
-        logger.error("Unable to find 'invalid token' details in the given message")
-        return None
-
-    _, details = invalid_token_error_msg.split(":")
-    line, column = details.split(",")
-    _, line_number = line.strip().split(" ")
-    _, column_number = column.strip().split(" ")
-
-    return XmlErrorLocation(int(line_number), int(column_number))
-
-
-def get_qgis_xml_error_context(
-    invalid_token_error_msg: str, fh: io.BufferedReader
-) -> Optional[tuple[str, str, str]]:
-    """Get a slice of the line where the exception occurred, with all faulty occurrences sanitized."""
-    location = get_qgis_xml_error_location(invalid_token_error_msg)
-    if location:
-        substitute = "?"
-        fh.seek(0)
-        for cursor_pos, line in enumerate(fh, start=1):
-            if location.line == cursor_pos:
-                faulty_char = line[location.column]
-                suffix_slice = line[: location.column - 1]
-                clean_safe_slice = suffix_slice.decode("utf-8").strip() + substitute
-
-                return (
-                    f"Unable to parse character: {repr(faulty_char)}",
-                    f"Replaced by '{substitute}' on line {location.line} that starts with:",
-                    html.escape(clean_safe_slice),
-                )
-
-    return None
 
 
 def check_valid_project_file(project_filename: Path) -> None:
